Some dashboard css tweaks
Removed the white line that was showing up under the recent recipes images.

Fixed the issue that caused the background image to zoom in on screens with a min-width of 500px.

Adjusted the buttons to reveal more of the bottom where it was getting cut off before.

Increased the font size of the recipe names on the recently saved.

Increased the bottom margin of the recently saved cards to make sure they weren't touching when on screens of 500px width.

Increased the description box to display at least two lines of a description.